Update the docker run example
I'm now specifying a container name, which is much more pleasant than
using container hashes to manage the system. Also name the extensions
volume anonymous by only refering to it from within the dockerfile; this
allows multiple containers based on the roon-extension-base image to run
at the same time, but with separate containers. It never needed to be
externally managed anyway.