Update readme with docker hub details
README.md: Now that this is up and running for me, update the instructions on how to use the pre-built images, and fix up the access to the USB device so that it actually works. It seems you need to use --device not simply pass through all of /dev/bus/usb as a volume mount.